Clear out ml2_vlan_allocations if CVX has no VLAN pool configured
If there is no vlan pool configured for a region, but CVX was
reachable, the arista_vlan type driver would fail to start.
We should treat this case as there being no VLANs assigned
to the region and clear out all unallocated vlans from the db.

+++ networking_arista/ml2/drivers/driver_helpers.py
@@ -125,13 +125,15 @@ class VlanSyncService(object):
                     'update'))
 
             for alloc in allocs:
+
                 if alloc.physical_network != 'default':
                     session.delete(alloc)
 
                 try:
                     assigned_vlans.remove(alloc.vlan_id)
                 except KeyError:
-                    session.delete(alloc)
+                    if not alloc.allocated:
+                       session.delete(alloc)
                     continue
 
                 if alloc.allocated and alloc.vlan_id in available_vlans:
